:sparkles: Feat: creating docker compose file for redis db and modularizing the application
