+++ docker-compose.yaml
@@ -1,0 +1,15 @@
+version: '3.9'
+
+services:
+
+  redis:
+    image: redis:latest
+    ports:
+      - "6379:6379"
+    volumes:
+      - redis:/data
+    restart: always
+
+volumes:
+  redis:
+    driver: local
