+++ internal/api-project/api.go
@@ -1,0 +1,30 @@
+package apiproject
+
+import (
+	"net/http"
+
+	"github.com/go-chi/chi/v5"
+	"github.com/go-chi/chi/v5/middleware"
+)
+
+type apiRespose struct {
+	Error string `json:"error,omitempty"`
+	Data  any    `json:"data,omitempty"`
+}
+
+func NewHandler(db map[string]string) http.Handler {
+	r := chi.NewMux()
+
+	r.Use(middleware.Recoverer)
+	r.Use(middleware.RequestID)
+	r.Use(middleware.Logger)
+
+	r.Route("/api", func(r chi.Router) {
+		r.Route("/url", func(r chi.Router) {
+			r.Post("/shorten", handleGetShortenendURL(db))
+			r.Get("/{code}", handleShortenURL(db))
+		})
+	})
+
+	return r
+}
